+++ satellite/src/main/java/com/linbit/linstor/core/devmgr/DeviceHandler.java
@@ -101,4 +101,13 @@ public interface DeviceHandler
 
     void closeAfterClone(VlmProviderObject<?> vlmDataRef, @Nullable String targetRscNameRef) throws StorageException;
 
+    /**
+     * Recursively calls all child layers processAfterClone method that should make any post clone processing needed.
+     * @param vlmSrcDataRef
+     * @param vlmTgtDataRef
+     * @param clonedPath
+     * @throws StorageException
+     */
+    void processAfterClone(VlmProviderObject<?> vlmSrcDataRef, VlmProviderObject<?> vlmTgtDataRef, String clonedPath)
+        throws StorageException;
 }